swap_roots
